+++ packages/@granit/identity/src/types/identity-user.ts
@@ -9,7 +9,7 @@ export type IdentityUser = {
   readonly firstName: string | null;
   readonly lastName: string | null;
   readonly enabled: boolean;
-  readonly extraProperties: Readonly<Record<string, string>>;
+  readonly metadata: Readonly<Record<string, string>>;
 };
 
 export type IdentityUserListParams = PaginationParams & {